# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "HTTP Basic with the subscription token as the password. The username is ignored — tokens are self-identifying. This is the form `git clone http://user:\u003ctoken\u003e@host/git/repo` produces, so git-over-HTTP works out of the box.",
        "title": "GiGot API",
        "contact": {
            "name": "Peter van de Pas"
        },
        "license": {
            "name": "MIT"
        },
        "version": "0.1.0"
    },
    "host": "localhost:3417",
    "basePath": "/api",
    "paths": {
...
        "/admin/repos/{name}/formidable": {
            "post": {
                "security": [
                    {
                        "SessionAuth": []
                    }
                ],
                "description": "Stamps .formidable/context.json on top of HEAD so the\nrepo picks up structured record-merge behaviour on subsequent\nwrites. Gated to server.formidable_first=true so generic-mode\noperators don't trip this accidentally. Idempotent: a repo\nthat already carries a valid marker returns stamped=false\nand writes no commit. On a successful stamp the server\nappends one `repo_convert_formidable` audit entry.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "formidable",
                    "admin"
                ],
                "summary": "Convert a plain repo to a Formidable context (admin only)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repo name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ConvertFormidableResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Server not in formidable_first mode",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Repo not found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Empty repo — nothing to stamp on top of",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/repos": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET lists all repositories, POST creates a new one. Set\nscaffold_formidable: true to seed the fresh repo with a\nstarter Formidable context (README, templates/basic.yaml,\nstorage/.gitkeep, and the .formidable/context.json marker)\nin an initial commit. POST appends one `repo_create` entry\nto refs/audit/main — see docs/design/audit-trail.md.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repos"
                ],
                "summary": "List or create repositories",
                "parameters": [
                    {
                        "description": "Create-repo body (POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.CreateRepoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.RepoListResponse"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET lists all repositories, POST creates a new one. Set\nscaffold_formidable: true to seed the fresh repo with a\nstarter Formidable context (README, templates/basic.yaml,\nstorage/.gitkeep, and the .formidable/context.json marker)\nin an initial commit. POST appends one `repo_create` entry\nto refs/audit/main — see docs/design/audit-trail.md.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repos"
                ],
                "summary": "List or create repositories",
                "parameters": [
                    {
                        "description": "Create-repo body (POST)",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/server.CreateRepoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.RepoListResponse"
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/server.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/repos/{name}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET returns repository details, DELETE removes it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repos"
                ],
                "summary": "Get or delete a repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.RepoInfo"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Missing or invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "GET returns repository details, DELETE removes it",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repos"
                ],
                "summary": "Get or delete a repository",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.RepoInfo"
                        }
                    },
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Missing or invalid bearer token",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/repos/{name}/formidable": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "One-call read of the Formidable-specific shape of a\nrepo: scaffold marker (version/by/at), templates list\nunder templates/, and storage layout under storage/.\nNon-Formidable repos return 200 with marker_present\nfalse and empty arrays so clients can distinguish\n\"not Formidable\" from \"doesn't exist\" or \"no access.\"",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "formidable"
                ],
                "summary": "Formidable-shape bootstrap for a repo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Repository name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.RepoFormidableResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    },
                    "405": {
                        "description": "Method Not Allowed",
                        "schema": {
                            "$ref": "#/definitions/server.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "server.ConvertFormidableResponse": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "repo": {
                    "$ref": "#/definitions/server.RepoInfo"
                },
                "stamped": {
                    "type": "boolean"
                }
            }
        },
...
        "server.CreateRepoRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "my-templates"
                },
                "scaffold_formidable": {
                    "type": "boolean",
                    "example": false
                },
                "source_url": {
                    "type": "string",
                    "example": "https://github.com/owner/repo.git"
                }
            }
        },
...
        "server.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "repository not found"
                }
            }
        },
        "server.FormidableMarkerView": {
            "type": "object",
            "properties": {
                "scaffolded_at": {
                    "type": "string"
                },
                "scaffolded_by": {
                    "type": "string"
                },
                "version": {
                    "type": "integer"
                }
            }
        },
        "server.FormidableStorage": {
            "type": "object",
            "properties": {
                "files": {
                    "type": "integer"
                },
                "template": {
                    "type": "string"
                }
            }
        },
        "server.FormidableTemplate": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                }
            }
        },
...
        "server.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "repository created"
                }
            }
        },
...
        "server.RepoFormidableResponse": {
            "type": "object",
            "properties": {
                "marker": {
                    "$ref": "#/definitions/server.FormidableMarkerView"
                },
                "marker_present": {
                    "type": "boolean"
                },
                "storage": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.FormidableStorage"
                    }
                },
                "templates": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.FormidableTemplate"
                    }
                }
            }
        },
        "server.RepoInfo": {
            "type": "object",
            "properties": {
                "commits": {
                    "type": "integer",
                    "example": 42
                },
                "default_branch": {
                    "type": "string",
                    "example": "main"
                },
                "destination_count": {
                    "type": "integer",
                    "example": 0
                },
                "empty": {
                    "type": "boolean",
                    "example": false
                },
                "has_formidable": {
                    "type": "boolean",
                    "example": false
                },
                "head": {
                    "type": "string",
                    "example": "a1b2c3d4e5f6"
                },
                "name": {
                    "type": "string",
                    "example": "my-templates"
                },
                "path": {
                    "type": "string",
                    "example": "repos/my-templates.git"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "server.RepoListResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer",
                    "example": 2
                },
                "repos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.RepoInfo"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "BearerAuth": {
            "description": "Enter your bearer token as: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie minted by POST /api/admin/login (or any successful OAuth callback). Required by every /api/admin/* endpoint and the /fragments/* template server. Browsers attach it automatically; programmatic callers must include the cookie header on every request.",
            "type": "apiKey",
            "name": "gigot_session",
            "in": "cookie"
        }
    }
}